removed separate ProjectList class and moved to validator.
