+++ src/main/java/io/openshift/launchpad/ui/booster/MissionControlValidator.java
@@ -40,6 +40,14 @@ public class MissionControlValidator
       });
    }
 
+   public List<String> getProjects(UIContext context)
+   {
+      Map<Object, Object> attributeMap = context.getAttributeMap();
+      List<String> authList = (List<String>) attributeMap.get(HttpHeaders.AUTHORIZATION);
+      String authHeader = (authList == null || authList.isEmpty()) ? null : authList.get(0);
+      return missionControlFacade.getProjects(authHeader);
+   }
+
    public boolean validateGitHubTokenExists(UIValidationContext context)
    {
       Map<Object, Object> attributeMap = context.getUIContext().getAttributeMap();
